Feat: Build input components
I have build the reusable dynamic input component, I added main props to it and the variables that can be sent as props are the following:
{ inputType="text", handleChange=(event)=>{console.log("An input is without handler function!!!")}, inputValue="",inputClasses="",inputName="",inputId="",inputPlaceholder="" }.
as you can see the default value if not sent for any input is going to be of type text, moreover adding change handler function for input is mandatory and we all should use the controlled form concept.
so if you will not add change handler function, the component will notify you by typing An input is without handler function!!! to remind the developer to add it.