Build: Add and configure React Router
I have added the react router dependency and added the Switch, so if there is an internal link, just Import the Link from react-router-dom and use it as follows
<Link to="path">some text</Link>.
Note*: when pushing you don't need to push the your changes on App.js just push your container and its components.
so if you made changes to App.js during work, thats fine, but dont' push that change to me as i will be adding them to the App.js.