Feat: Build and configure google signin and Signup
I have applied so many changes the main thing was configuring the firebase. Then I have created google sign-in using firebase auth.
After that I created the form that is used to get the additional information to be added to user's account in the database.
A state which will store the login session was created. Navbar will need some edits as well i will have to create an issue for that.